Refactor pastebin availability routine
